Add files via upload
This may be helpful to anyone attempting to to automate outside lights using motion detection and auto on/off based on location's sunrise/sunset. It will turn on outdoor lights at 30% when the sun is down. If motion is detected it will change the brightness to 100% for 5 minutes after the last motion event is detected.